reorganised the schemes' package structure
