--- ckks/keygen.go
@@ -1,8 +1,0 @@
-package ckks
-
-import "github.com/ldsec/lattigo/v2/rlwe"
-
-// NewKeyGenerator creates a rlwe.KeyGenerator instance from the CKKS parameters.
-func NewKeyGenerator(params Parameters) rlwe.KeyGenerator {
-	return rlwe.NewKeyGenerator(params.Parameters)
-}
