+++ ckks/keys.go
@@ -2,6 +2,11 @@ package ckks
 
 import "github.com/ldsec/lattigo/v2/rlwe"
 
+// NewKeyGenerator creates a rlwe.KeyGenerator instance from the CKKS parameters.
+func NewKeyGenerator(params Parameters) rlwe.KeyGenerator {
+	return rlwe.NewKeyGenerator(params.Parameters)
+}
+
 // BootstrappingKey is a type for a CKKS bootstrapping key, wich regroups the necessary public relinearization
 // and rotation keys (i.e., an EvaluationKey).
 type BootstrappingKey rlwe.EvaluationKey
