+++ dckks/keyswitch.go
@@ -22,3 +22,20 @@ func (cks *CKSProtocol) KeySwitchCKKS(combined *drlwe.CKSShare, ct *ckks.Ciphert
 	ctOut.Scale = ct.Scale
 	cks.CKSProtocol.KeySwitch(combined, ct.Ciphertext, ctOut.Ciphertext)
 }
+
+// PCKSProtocol is the structure storing the parameters for the collective public key-switching.
+type PCKSProtocol struct {
+	drlwe.PCKSProtocol
+}
+
+// NewPCKSProtocol creates a new PCKSProtocol object and will be used to re-encrypt a ciphertext ctx encrypted under a secret-shared key mong j parties under a new
+// collective public-key.
+func NewPCKSProtocol(params ckks.Parameters, sigmaSmudging float64) *PCKSProtocol {
+	return &PCKSProtocol{*drlwe.NewPCKSProtocol(params.Parameters, sigmaSmudging)}
+}
+
+// KeySwitchCKKS performs the actual keyswitching operation on a ciphertext ct and put the result in ctOut
+func (pcks *PCKSProtocol) KeySwitchCKKS(combined *drlwe.PCKSShare, ct, ctOut *ckks.Ciphertext) {
+	pcks.PCKSProtocol.KeySwitch(combined, ct.Ciphertext, ctOut.Ciphertext)
+	ctOut.Scale = ct.Scale
+}
