--- dckks/rotkey_gen.go
@@ -1,16 +1,0 @@
-package dckks
-
-import (
-	"github.com/ldsec/lattigo/v2/ckks"
-	"github.com/ldsec/lattigo/v2/drlwe"
-)
-
-// RTGProtocol is the structure storing the parameters for the collective rotation-keys generation.
-type RTGProtocol struct {
-	drlwe.RTGProtocol
-}
-
-// NewRotKGProtocol creates a new rotkg object and will be used to generate collective rotation-keys from a shared secret-key among j parties.
-func NewRotKGProtocol(params ckks.Parameters) (rtg *RTGProtocol) {
-	return &RTGProtocol{*drlwe.NewRTGProtocol(params.Parameters)}
-}
